requirements.txt
dash==1.20.0
opencv-python==4.5.1.48
Pillow==8.2.0
numpy==1.20.0
gunicorn==20.1.0